test: move the File/S3File GC accounting tests to blob.test.ts
heap-snapshot.test.ts has two pre-existing tests (URL heap size, Headers)
that take 12-17 s under a debug+ASAN build, past the default per-test
timeout, so the file as a whole does not pass under bun bd regardless of
this change. The new tests sit next to the File constructor tests instead.
