+++ test/js/bun/util/heap-snapshot.test.ts
@@ -32,124 +32,6 @@ describe("Native types report their size correctly", () => {
     delete globalThis.formData;
   });
 
-  it("File", () => {
-    const payload = Buffer.alloc(1024 * 1024, "abc");
-    const name = "file.bin";
-    const blobSize = estimateShallowMemoryUsageOf(new Blob([payload]));
-    expect(blobSize).toBeGreaterThanOrEqual(payload.byteLength);
-
-    // A File is a Blob plus a name, so it reports at least what the same Blob does.
-    const file = new File([payload], name);
-    expect(estimateShallowMemoryUsageOf(file)).toBeGreaterThanOrEqual(blobSize + name.length);
-
-    const type = "application/x-custom-type";
-    const typedBlobSize = estimateShallowMemoryUsageOf(new Blob([payload], { type }));
-    expect(estimateShallowMemoryUsageOf(new File([payload], name, { type }))).toBeGreaterThanOrEqual(
-      typedBlobSize + name.length,
-    );
-
-    // No bytes, only a name.
-    const longName = Buffer.alloc(64 * 1024, "n").toString();
-    expect(estimateShallowMemoryUsageOf(new File([], longName))).toBeGreaterThanOrEqual(longName.length);
-
-    // A single Blob part shares the source's store instead of copying it; the
-    // new File still has to report it (and not just repeat what the source reported).
-    expect(estimateShallowMemoryUsageOf(new File([file], name))).toBeGreaterThanOrEqual(payload.byteLength);
-
-    // Over a file-backed or S3-backed store the name lives on the File itself
-    // instead of in the store.
-    expect(estimateShallowMemoryUsageOf(new File([Bun.file(import.meta.path)], longName))).toBeGreaterThanOrEqual(
-      longName.length,
-    );
-    const s3 = new Bun.S3Client({ accessKeyId: "id", secretAccessKey: "secret", bucket: "bucket" }).file("key");
-    expect(estimateShallowMemoryUsageOf(new File([s3], longName))).toBeGreaterThanOrEqual(longName.length);
-
-    class MyFile extends File {}
-    const subclassed = new MyFile([payload], name);
-    expect(subclassed).toBeInstanceOf(MyFile);
-    expect(estimateShallowMemoryUsageOf(subclassed)).toBeGreaterThanOrEqual(payload.byteLength);
-
-    // FormData entries carry the File's reported size, like the Blob entries above.
-    const formData = new FormData();
-    const emptyFormDataSize = estimateShallowMemoryUsageOf(formData);
-    formData.append("file", file);
-    expect(estimateShallowMemoryUsageOf(formData)).toBeGreaterThanOrEqual(emptyFormDataSize + payload.byteLength);
-    expect(estimateShallowMemoryUsageOf(formData.get("file") as File)).toBeGreaterThanOrEqual(payload.byteLength);
-  });
-
-  // `new File()` and S3 files are the two Blob wrappers created outside the
-  // generated Blob constructor. Bun lowers JSC's minimum allocation budget per GC
-  // cycle (largeHeapSize) to 8 MiB, which is what a fresh process with almost
-  // nothing live gets, so 512 instances that each own ~64 KB (a payload, or the
-  // credentials an S3 file copies) are worth several collections, but only if the
-  // wrapper reports that size when it is created: the cells themselves are a few
-  // dozen bytes each and never get near the budget. `className` is the heap type
-  // both the prototype and the instances are counted under.
-  describe.concurrent.each([
-    {
-      className: "Blob",
-      setup: /* js */ `const payload = Buffer.alloc(64 * 1024, "abc");`,
-      construct: /* js */ `new File([payload], "file.bin")`,
-    },
-    {
-      className: "S3File",
-      setup: /* js */ `
-        const client = new Bun.S3Client({
-          accessKeyId: "id",
-          secretAccessKey: "secret",
-          endpoint: "http://localhost:1",
-          bucket: Buffer.alloc(64 * 1024, "b").toString(),
-        });
-      `,
-      construct: /* js */ `client.file("key")`,
-    },
-  ])("$construct", ({ className, setup, construct }) => {
-    it("reports its size to the GC when constructed", async () => {
-      await using proc = Bun.spawn({
-        cmd: [
-          bunExe(),
-          "-e",
-          /* js */ `
-            const { heapStats } = require("bun:jsc");
-            ${setup}
-            const count = () => heapStats().objectTypeCounts[${JSON.stringify(className)}];
-            // Stays referenced: creating it also creates the lazily allocated prototype.
-            const warmup = ${construct};
-            // heapStats() runs a full collection itself if none has happened yet
-            // (BUN_GC_TIMER_DISABLE=1 skips the startup one). Taking the baseline
-            // here absorbs that, so the counts below are plain live-cell counts.
-            const before = count();
-            const held = ${construct};
-            const withHeld = count();
-            const created = 512;
-            for (let i = 0; i < created; i++) {
-              ${construct};
-            }
-            const alive = count() - withHeld;
-            console.log(JSON.stringify({ before, withHeld, created, alive, kept: [warmup.name, held.name] }));
-          `,
-        ],
-        env: bunEnv,
-        stderr: "pipe",
-      });
-      const [stdout, stderr, exitCode] = await Promise.all([proc.stdout.text(), proc.stderr.text(), proc.exited]);
-      const stderrLines = stderr
-        .split("\n")
-        .filter(line => line && !line.startsWith("WARNING: ASAN interferes"))
-        .join("\n");
-      expect(stderrLines).toBe("");
-      const { before, withHeld, created, alive } = JSON.parse(stdout);
-      // Positive control: the counter really counts these instances.
-      expect(withHeld).toBe(before + 1);
-      // Seen: 10-25 alive when the size is reported on construction; roughly one
-      // budget's worth (~125 of these) is what can pile up after the last
-      // collection, so the budget would have to double before the bound is at
-      // risk. Exactly `created` survive when the size is not reported.
-      expect(alive).toBeLessThan(created / 2);
-      expect(exitCode).toBe(0);
-    });
-  });
-
   it("Request", () => {
     var request = new Request("https://example.com", {
       body: Buffer.alloc(1024 * 1024 * 2, "yoo"),
